+++ android/src/main/java/hamza/dali/flutter_osm_plugin/Constants.java
@@ -1,4 +1,19 @@
 package hamza.dali.flutter_osm_plugin;
 
+import android.graphics.PorterDuff;
+import android.graphics.drawable.BitmapDrawable;
+import android.graphics.drawable.Drawable;
+import android.os.Build;
+
+import androidx.core.content.ContextCompat;
+
 public class Constants {
+
+    public static final String STARTPOSITIONROAD="START";
+    public static final String MIDDLEPOSITIONROAD="MIDDLE";
+    public static final String ENDPOSITIONROAD="END";
+
+
+
+
 }