+++ android/src/main/java/hamza/dali/flutter_osm_plugin/FlutterOsmView.java
@@ -1,10 +1,9 @@
 package hamza.dali.flutter_osm_plugin;
 
-import android.Manifest;
+
 import android.app.Activity;
 import android.app.Application;
 import android.content.Context;
-import android.content.pm.PackageManager;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.graphics.Color;
@@ -17,7 +16,6 @@ import android.os.Bundle;
 import android.os.Handler;
 import android.preference.PreferenceManager;
 import android.util.Log;
-import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.LinearLayout;
@@ -34,30 +32,22 @@ import org.osmdroid.bonuspack.routing.Road;
 import org.osmdroid.bonuspack.routing.RoadManager;
 import org.osmdroid.config.Configuration;
 import org.osmdroid.events.MapEventsReceiver;
-import org.osmdroid.events.MapListener;
-import org.osmdroid.events.ScrollEvent;
-import org.osmdroid.events.ZoomEvent;
-import org.osmdroid.tileprovider.MapTileProviderBase;
 import org.osmdroid.tileprovider.tilesource.TileSourceFactory;
 import org.osmdroid.util.GeoPoint;
 import org.osmdroid.views.CustomZoomButtonsController;
 import org.osmdroid.views.MapView;
 import org.osmdroid.views.overlay.MapEventsOverlay;
 import org.osmdroid.views.overlay.Marker;
+import org.osmdroid.views.overlay.Overlay;
 import org.osmdroid.views.overlay.Polyline;
 import org.osmdroid.views.overlay.mylocation.GpsMyLocationProvider;
 import org.osmdroid.views.overlay.mylocation.MyLocationNewOverlay;
 
-import java.io.ByteArrayInputStream;
-import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import io.flutter.embedding.android.FlutterView;
-import io.flutter.embedding.engine.FlutterEngine;
-import io.flutter.embedding.engine.plugins.activity.ActivityAware;
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
 import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.EventChannel;
@@ -74,6 +64,13 @@ import static hamza.dali.flutter_osm_plugin.FlutterOsmPlugin.STARTED;
 import static hamza.dali.flutter_osm_plugin.FlutterOsmPlugin.STOPPED;
 import static io.flutter.plugin.common.MethodChannel.Result;
 
+
+enum PositionMarker {
+    START,
+    MIDDLE,
+    END;
+}
+
 public class FlutterOsmView implements
         Application.ActivityLifecycleCallbacks,
         DefaultLifecycleObserver,
@@ -81,7 +78,7 @@ public class FlutterOsmView implements
         PlatformView,
         EventChannel.StreamHandler,
         MethodChannel.MethodCallHandler {
-    MapView map;
+    private MapView map;
     private MyLocationNewOverlay locationNewOverlay;
     private Context context;
     private final MethodChannel methodChannel;
@@ -92,12 +89,13 @@ public class FlutterOsmView implements
     private final AtomicInteger activityState;
     private PluginRegistry.Registrar registrar;
     private Bitmap customMarkerIcon;
+    private HashMap<String, Bitmap> customRoadMarkerIcon = new HashMap<>();
     private MapEventsOverlay mapEventsOverlay;
     private OSRMRoadManager roadManager;
-
+    private Integer roadColor = null;
     private final double defaultZoom = 10.;
-    Handler handler;
-    Runnable runnable;
+    private final String url = "router.project-osrm.org/route/v1/driving/";
+    private boolean useScureURL = true;
 
     public FlutterOsmView(Context ctx,
                           PluginRegistry.Registrar registrar,
@@ -160,27 +158,38 @@ public class FlutterOsmView implements
 
     private void initPosition(MethodCall methodCall, Result result) {
 
-        HashMap<String, Double> args = (HashMap) methodCall.arguments;
+        HashMap<String, Double> args = (HashMap<String, Double>) methodCall.arguments;
         map.getOverlays().clear();
         GeoPoint geoPoint = new GeoPoint(args.get("lat"), args.get("lon"));
-        addMarker(geoPoint, defaultZoom,null);
+        addMarker(geoPoint, defaultZoom, null, null);
         result.success(null);
     }
 
-    void addMarker(GeoPoint geoPoint, double zoom,@Nullable  Integer color) {
+    private void addMarker(GeoPoint geoPoint, double zoom, @Nullable Integer color, @Nullable PositionMarker Posmarker) {
         Marker marker = new Marker(map);
-        Drawable iconDrawable;
-        if (customMarkerIcon != null) {
-            iconDrawable = new BitmapDrawable(getActivity().getResources(), customMarkerIcon);
-            if(color!=null)
-            iconDrawable.setColorFilter(color, PorterDuff.Mode.SRC_OVER);
-        } else {
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-                iconDrawable = getActivity().getDrawable(R.drawable.ic_location_on_red_24dp);
-            } else {
-                iconDrawable = ContextCompat.getDrawable(getActivity(), R.drawable.ic_location_on_red_24dp);
+        Drawable iconDrawable = null;
+        if (Posmarker != null) {
+            switch (Posmarker) {
+                case START:
+                    if (customRoadMarkerIcon.containsKey(Constants.STARTPOSITIONROAD)) {
+                        iconDrawable = new BitmapDrawable(getActivity().getResources(), customRoadMarkerIcon.get(Constants.STARTPOSITIONROAD));
+                    }
+                    break;
+                case MIDDLE:
+                    if (customRoadMarkerIcon.containsKey(Constants.MIDDLEPOSITIONROAD)) {
+                        iconDrawable = new BitmapDrawable(getActivity().getResources(), customRoadMarkerIcon.get(Constants.MIDDLEPOSITIONROAD));
+                    }
+                    break;
+                case END:
+                    if (customRoadMarkerIcon.containsKey(Constants.ENDPOSITIONROAD)) {
+                        iconDrawable = new BitmapDrawable(getActivity().getResources(), customRoadMarkerIcon.get(Constants.ENDPOSITIONROAD));
+                    }
+                    break;
             }
         }
+        if (iconDrawable == null)
+            iconDrawable = getDefaultIconDrawable(color);
+
 
         marker.setIcon(iconDrawable);
 
@@ -188,6 +197,22 @@ public class FlutterOsmView implements
         map.getController().setZoom(10.);
         map.getController().animateTo(geoPoint);
         map.getOverlays().add(marker);
+    }
+
+    private Drawable getDefaultIconDrawable(@Nullable Integer color) {
+        Drawable iconDrawable = null;
+        if (customMarkerIcon != null) {
+            iconDrawable = new BitmapDrawable(getActivity().getResources(), customMarkerIcon);
+            if (color != null)
+                iconDrawable.setColorFilter(color, PorterDuff.Mode.SRC_OVER);
+        } else {
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+                iconDrawable = getActivity().getDrawable(R.drawable.ic_location_on_red_24dp);
+            } else {
+                iconDrawable = ContextCompat.getDrawable(getActivity(), R.drawable.ic_location_on_red_24dp);
+            }
+        }
+        return iconDrawable;
     }
 
     private void enableMyLocation(Result result) {
@@ -231,6 +256,8 @@ public class FlutterOsmView implements
     @Override
     public void onMethodCall(MethodCall call, MethodChannel.Result result) {
         switch (call.method) {
+            case "use#secure":
+                setSecureURL(call,result);
             case "Zoom":
                 setZoom(call, result);
                 break;
@@ -258,55 +285,72 @@ public class FlutterOsmView implements
             case "marker#icon":
                 changeIcon(call, result);
                 break;
+            case "road#color":
+                setRoadColor(call, result);
+                break;
+            case "road#markers":
+                setMarkerRoad(call, result);
+                break;
             default:
                 result.notImplemented();
         }
 
     }
 
-    private void drawRoad(MethodCall call,final Result result) {
+    private void setSecureURL(MethodCall call, Result result) {
+        useScureURL=(boolean) call.arguments;
+        result.success(null);
+    }
+
+    private void drawRoad(MethodCall call, final Result result) {
         final List<HashMap<String, Double>> points = (List<HashMap<String, Double>>) call.arguments;
 
         if (roadManager == null) {
             roadManager = new OSRMRoadManager(getApplication());
-            roadManager.setService("https://router.project-osrm.org/route/v1/driving/");
+            if(useScureURL)
+            roadManager.setService("https://" + url);
         }
-        for(int i=0;i<map.getOverlays().size();i++){
-            if(map.getOverlays().get(i).getClass() == Polyline.class){
+        for (int i = 0; i < map.getOverlays().size(); i++) {
+            if (map.getOverlays().get(i).getClass() == Polyline.class) {
                 map.getOverlays().remove(i);
             }
         }
         map.invalidate();
-        if (handler != null) {
-            handler.removeCallbacks(runnable);
-        }
-        new AsyncTask<List<HashMap<String,Double>>,Void,Road>() {
+
+        new AsyncTask<List<HashMap<String, Double>>, Void, Road>() {
             ArrayList<GeoPoint> waypoints = new ArrayList();
+
             @Override
             protected Road doInBackground(List<HashMap<String, Double>>... lists) {
 
                 for (HashMap<String, Double> hashmap : lists[0]) {
                     waypoints.add(new GeoPoint(hashmap.get("lat"), hashmap.get("lon")));
                 }
-                Road road = roadManager.getRoad(waypoints);
-
-                return road;
+                return roadManager.getRoad(waypoints);
             }
 
             @Override
             protected void onPostExecute(Road road) {
                 super.onPostExecute(road);
-                if(getActivity()!=null){
-                    if(road.mRouteHigh.size()>2){
+                if (getActivity() != null) {
+                    if (road.mRouteHigh.size() > 2) {
+                        for (Overlay overlay : map.getOverlays()) {
+                            if (overlay instanceof Marker || overlay instanceof Polyline) {
+                                map.getOverlays().remove(overlay);
+                            }
+                        }
                         Polyline roadOverlay = RoadManager.buildRoadOverlay(road);
                         roadOverlay.getOutlinePaint().setColor(Color.GREEN);
-                        addMarker(waypoints.get(waypoints.size()-1),14,Color.RED);
-                        addMarker(waypoints.get(0),14,Color.RED);
+                        if (roadColor != null) {
+                            roadOverlay.getOutlinePaint().setColor(roadColor);
+                        }
+                        addMarker(waypoints.get(0), 14, Color.RED, PositionMarker.START);
+                        addMarker(waypoints.get(waypoints.size() - 1), 14, Color.RED, PositionMarker.END);
                         map.getOverlays().add(roadOverlay);
                         map.invalidate();
                         result.success(null);
-                    }else{
-                        result.error("423","Opps!we cannot draw road correctly!",null);
+                    } else {
+                        result.error("423", "Opps!we cannot draw road correctly!", null);
                     }
 
                 }
@@ -332,7 +376,7 @@ public class FlutterOsmView implements
                     Hashmap.put("lat", p.getLatitude());
                     Hashmap.put("lon", p.getLongitude());
                     map.getOverlays().remove(0);
-                    addMarker(p, 15.,null);
+                    addMarker(p, 15., null, null);
                     mapEventsOverlay = null;
                     _result.success(Hashmap);
                     return false;
@@ -346,7 +390,7 @@ public class FlutterOsmView implements
     private void changeIcon(MethodCall call, Result result) {
         try {
             byte[] bytes = (byte[]) call.arguments;
-            Log.d("bytes", "" + call.arguments.toString());
+
             customMarkerIcon = BitmapFactory.decodeByteArray(bytes, 0, bytes.length);
             //customMarkerIcon.recycle();
             result.success(null);
@@ -357,6 +401,41 @@ public class FlutterOsmView implements
         }
 
     }
+
+    private void setRoadColor(MethodCall call, Result result) {
+        List<Integer> argb = (List<Integer>) call.arguments;
+        roadColor = Color.rgb(argb.get(0), argb.get(1), argb.get(2));
+        result.success(null);
+    }
+
+    private void setMarkerRoad(MethodCall call, Result result) {
+        try {
+            HashMap<String, byte[]> icons = (HashMap<String, byte[]>) call.arguments;
+            if (icons.containsKey(Constants.STARTPOSITIONROAD)) {
+                customRoadMarkerIcon.put(Constants.STARTPOSITIONROAD,
+                        BitmapFactory.decodeByteArray(icons.get(Constants.STARTPOSITIONROAD), 0, icons.get(Constants.STARTPOSITIONROAD).length));
+            }
+            if (icons.containsKey(Constants.ENDPOSITIONROAD)) {
+                customRoadMarkerIcon.put(Constants.ENDPOSITIONROAD,
+                        BitmapFactory.decodeByteArray(icons.get(Constants.ENDPOSITIONROAD), 0,
+                                icons.get(Constants.ENDPOSITIONROAD).length));
+            }
+            if (icons.containsKey(Constants.MIDDLEPOSITIONROAD)) {
+                customRoadMarkerIcon.put(Constants.MIDDLEPOSITIONROAD,
+                        BitmapFactory.decodeByteArray(icons.get(Constants.MIDDLEPOSITIONROAD), 0,
+                                icons.get(Constants.MIDDLEPOSITIONROAD).length));
+            }
+            //customMarkerIcon = BitmapFactory.decodeByteArray(bytes, 0, bytes.length);
+            //customMarkerIcon.recycle();
+            result.success(null);
+        } catch (Exception e) {
+            Log.d("err", e.getMessage());
+            customMarkerIcon = null;
+            result.error("500", "Cannot make markerIcon custom", "");
+        }
+
+    }
+
 
     private void userPosition(MethodCall call, Result result) {
         if (this.locationNewOverlay == null || !this.locationNewOverlay.isMyLocationEnabled()) {
@@ -425,10 +504,7 @@ public class FlutterOsmView implements
     @Override
     public void onActivityPaused(Activity activity) {
         map.onPause();
-        if (handler != null) {
-            handler.removeCallbacks(runnable);
-            runnable=null;
-        }
+
     }
 
     @Override