+++ lib/src/osm_flutter.dart
@@ -7,7 +7,9 @@ import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_osm_plugin/flutter_osm_plugin.dart';
 import 'package:flutter_osm_plugin/src/geo_point_exception.dart';
+import 'package:flutter_osm_plugin/src/job_alert_dialog.dart';
 import 'package:flutter_osm_plugin/src/marker.dart';
+import 'package:flutter_osm_plugin/src/road.dart';
 import 'package:flutter_osm_plugin/src/road_exception.dart';
 import 'package:location_permissions/location_permissions.dart';
 
@@ -17,6 +19,8 @@ class OSMFlutter extends StatefulWidget {
   final bool showZoomController;
   final GeoPoint initPosition;
   final MarkerIcon markerIcon;
+  final Road road;
+  final bool useSecureURL;
   OSMFlutter({
     Key key,
     this.currentLocation = true,
@@ -24,6 +28,8 @@ class OSMFlutter extends StatefulWidget {
     this.showZoomController = false,
     this.initPosition,
     this.markerIcon,
+    this.road,
+    this.useSecureURL = true,
   }) : super(key: key);
 
   static OSMFlutterState of<T>(BuildContext context, {bool nullOk = false}) {
@@ -50,13 +56,16 @@ class OSMFlutterState extends State<OSMFlutter> {
   PermissionStatus _permission;
   //_OsmCreatedCallback _osmCreatedCallback;
   _OsmController _osmController;
-  GlobalKey _key;
+  GlobalKey _key, _startIconKey, _endIconKey, _midddleIconKey;
 
   @override
   void initState() {
     super.initState();
     _key = GlobalKey();
-    Future.delayed(Duration(milliseconds: 200), () async {
+    _startIconKey = GlobalKey();
+    _endIconKey = GlobalKey();
+    _midddleIconKey = GlobalKey();
+    Future.delayed(Duration(milliseconds: 150), () async {
       //check location permission
       _permission = await LocationPermissions().checkPermissionStatus();
       if (_permission == PermissionStatus.denied) {
@@ -68,11 +77,36 @@ class OSMFlutterState extends State<OSMFlutter> {
       } else if (_permission == PermissionStatus.granted) {
         if (widget.currentLocation) await _checkServiceLocation();
       }
+
+      await this._osmController.setSecureURL(widget.useSecureURL);
+
       if (widget.markerIcon != null) {
         await changeIconMarker(_key);
       }
       if (widget.initPosition != null) {
         await changeLocation(widget.initPosition);
+      }
+      if (widget.road != null) {
+        await showDialog(
+            context: context,
+            barrierDismissible: false,
+            builder: (ctx) {
+              return JobAlertDialog(
+                callback: () async {
+                  await this._osmController.setColorRoad(
+                        widget.road.roadColor.red,
+                        widget.road.roadColor.green,
+                        widget.road.roadColor.blue,
+                      );
+                  await this._osmController.setMarkersRoad(
+                        _startIconKey,
+                        _endIconKey,
+                        _midddleIconKey,
+                      );
+                  Navigator.pop(ctx);
+                },
+              );
+            });
       }
     });
   }
@@ -83,9 +117,10 @@ class OSMFlutterState extends State<OSMFlutter> {
     this._osmController.addPosition(p);
   }
 
-Future changeIconMarker(GlobalKey key) async{
+  Future changeIconMarker(GlobalKey key) async {
     await this._osmController.customMarker(key);
   }
+
   ///zoom in/out
   /// positive value:zoomIN
   /// negative value:zoomOut
@@ -113,12 +148,16 @@ Future changeIconMarker(GlobalKey key) async{
     GeoPoint p = await this._osmController.pickLocation();
     return p;
   }
+
   //draw road
-  Future<void> drawRoad(GeoPoint start,GeoPoint end)async{
-      assert(start!=null&&end!=null,"you cannot make road without 2 point");
-      assert(start.latitude!=end.latitude || start.longitude!=end.longitude,"you cannot make road with same geopoint");
-      await this._osmController.drawRoad(start, end);
+  Future<void> drawRoad(GeoPoint start, GeoPoint end) async {
+    assert(
+        start != null && end != null, "you cannot make road without 2 point");
+    assert(start.latitude != end.latitude || start.longitude != end.longitude,
+        "you cannot make road with same geopoint");
+    await this._osmController.drawRoad(start, end);
   }
+
   Future<void> _checkServiceLocation() async {
     ServiceStatus serviceStatus =
         await LocationPermissions().checkServiceStatus();
@@ -163,10 +202,7 @@ Future changeIconMarker(GlobalKey key) async{
     if (defaultTargetPlatform == TargetPlatform.android) {
       return Stack(
         children: <Widget>[
-          RepaintBoundary(
-            key: _key,
-            child: widget.markerIcon,
-          ),
+          widgetConfigMap(),
           AndroidView(
             viewType: 'plugins.dali.hamza/osmview',
             onPlatformViewCreated: _onPlatformViewCreated,
@@ -177,6 +213,38 @@ Future changeIconMarker(GlobalKey key) async{
     }
     return Text(
         '$defaultTargetPlatform is not yet supported by the osm plugin');
+  }
+
+  Widget widgetConfigMap() {
+    return Positioned(
+      top: -100,
+      child: Stack(
+        children: <Widget>[
+          RepaintBoundary(
+            key: _key,
+            child: widget.markerIcon,
+          ),
+          if (widget.road?.endIcon != null) ...[
+            RepaintBoundary(
+              key: _endIconKey,
+              child: widget.road.endIcon,
+            ),
+          ],
+          if (widget.road?.startIcon != null) ...[
+            RepaintBoundary(
+              key: _startIconKey,
+              child: widget.road.startIcon,
+            ),
+          ],
+          if (widget.road?.middleIcon != null) ...[
+            RepaintBoundary(
+              key: _midddleIconKey,
+              child: widget.road.middleIcon,
+            ),
+          ],
+        ],
+      ),
+    );
   }
 
   void _onPlatformViewCreated(int id) {
@@ -191,6 +259,10 @@ class _OsmController {
 
   final MethodChannel _channel;
   //final EventChannel _eventChannel;
+
+  Future<void> setSecureURL(bool secure) async {
+    return await _channel.invokeMethod('use#secure', secure);
+  }
 
   Future<void> zoom(double zoom) async {
     assert(zoom != null);
@@ -217,14 +289,38 @@ class _OsmController {
           await _channel.invokeMapMethod("user#pickPosition", null);
       return GeoPoint(latitude: map["lat"], longitude: map["lon"]);
     } on PlatformException catch (e) {
-     throw GeoPointException(msg: e.message);
+      throw GeoPointException(msg: e.message);
     }
   }
 
   Future<void> customMarker(GlobalKey globalKey) async {
     Uint8List icon = await _capturePng(globalKey);
-    print(icon);
-    await _channel.invokeMapMethod("marker#icon", icon);
+    await _channel.invokeMethod("marker#icon", icon);
+  }
+
+  Future<void> setColorRoad(int r, int g, int b) async {
+    await _channel.invokeMethod("road#color", [r, g, b]);
+  }
+
+  Future<void> setMarkersRoad(
+      GlobalKey startKey, GlobalKey endKey, GlobalKey middleKey) async {
+    Map<String, Uint8List> bitmaps = {};
+    if (startKey.currentContext != null) {
+      print("start");
+      Uint8List marker = await _capturePng(startKey);
+      bitmaps.putIfAbsent("START", () => marker);
+    }
+    if (endKey.currentContext != null) {
+      print("end");
+      Uint8List marker = await _capturePng(endKey);
+      bitmaps.putIfAbsent("END", () => marker);
+    }
+    if (middleKey.currentContext != null) {
+      print("end");
+      Uint8List marker = await _capturePng(middleKey);
+      bitmaps.putIfAbsent("MIDDLE", () => marker);
+    }
+    await _channel.invokeMethod("road#markers", bitmaps);
   }
 
   Future<Uint8List> _capturePng(GlobalKey globalKey) async {
@@ -249,12 +345,12 @@ class _OsmController {
 
   ///draw road
   Future<void> drawRoad(GeoPoint p, GeoPoint p2) async {
-    try{
+    try {
       return await _channel.invokeMethod("road", [
-      {"lon": p.longitude, "lat": p.latitude},
-      {"lon": p2.longitude, "lat": p2.latitude}
-    ]);
-    }on PlatformException catch(e){
+        {"lon": p.longitude, "lat": p.latitude},
+        {"lon": p2.longitude, "lat": p2.latitude}
+      ]);
+    } on PlatformException catch (e) {
       throw RoadException(msg: e.message);
     }
   }